+++ src/components/plots/scatter.tsx
@@ -67,6 +67,8 @@ const Scatter = withTooltip<DotsProps, BasePoint>(
       return newConfig as CompleteScatterPlotOptions;
     }, [data, options]);
 
+    //const decimationThreshold = useMemo(() => (config.y.domain.max - config.y.domain.min)/25, [config])
+
     const checkBoundaries = useCallback(
       (d: BasePoint) => {
         return (
@@ -79,13 +81,8 @@ const Scatter = withTooltip<DotsProps, BasePoint>(
       [config]
     );
 
-    const xMax = useMemo(() => {
-      return width - defaultMargin.left - defaultMargin.right;
-    }, [width]);
-
-    const yMax = useMemo(() => {
-      return height - defaultMargin.top - defaultMargin.bottom;
-    }, [height]);
+    const xMax = useMemo(() => width - defaultMargin.left - defaultMargin.right, [width]);
+    const yMax = useMemo(() => height - defaultMargin.top - defaultMargin.bottom, [height]);
 
     const xScale = useMemo(
       () =>
@@ -161,9 +158,7 @@ const Scatter = withTooltip<DotsProps, BasePoint>(
     if (data.length === 0) {
       return (
         <VStack width={width} height={height}>
-          <Heading variant='notFound'>
-            No Data Available
-          </Heading>
+          <Heading variant='notFound'>No Data Available</Heading>
         </VStack>
       );
     }
@@ -179,6 +174,7 @@ const Scatter = withTooltip<DotsProps, BasePoint>(
             y={defaultMargin.top}
             fill='white'
             aria-label='Graph'
+            shapeRendering='optimizeSpeed'
             onMouseMove={handleMouseMove}
             onMouseLeave={handleMouseLeave}
             onTouchMove={handleMouseMove}
@@ -186,8 +182,8 @@ const Scatter = withTooltip<DotsProps, BasePoint>(
             onClick={handleMouseClick}
           />
           <Group pointerEvents='none' left={defaultMargin.left} top={defaultMargin.top}>
-            <GridRows scale={yScale} width={xMax} height={yMax} stroke='#e0e0e0' />
-            <GridColumns scale={xScale} width={xMax} height={yMax} stroke='#e0e0e0' />
+            <GridRows shapeRendering='optimizeSpeed' scale={yScale} width={xMax} height={yMax} stroke='#e0e0e0' />
+            <GridColumns shapeRendering='optimizeSpeed' scale={xScale} width={xMax} height={yMax} stroke='#e0e0e0' />
             <AxisBottom label={config.x.label} top={yMax} scale={xScale} numTicks={5} />
             <AxisLeft label={config.y.label} scale={yScale} numTicks={5} />
             {data.map(
@@ -199,6 +195,7 @@ const Scatter = withTooltip<DotsProps, BasePoint>(
                     className='dot'
                     cx={xScale(x(point))}
                     cy={yScale(y(point))}
+                    shapeRendering='optimizeSpeed'
                     r={config.points.dotRadius}
                     fill={tooltipData === point ? "pink" : "#ff5733"}
                   />